requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
attrs==19.1.0
Automat==0.7.0
backcall==0.1.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
constantly==15.1.0
cryptography==2.6.1
cssselect==1.0.3
decorator==4.4.0
Django==2.2.2
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
ipython==7.5.0
ipython-genutils==0.2.0
isort==4.3.20
jedi==0.13.3
lazy-object-proxy==1.4.1
lxml==4.3.3
mccabe==0.6.1
parsel==1.5.1
parso==0.4.0
pickleshare==0.7.5
Pillow==6.0.0
prompt-toolkit==2.0.9
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyDispatcher==2.0.5
Pygments==2.4.2
PyHamcrest==1.9.0
pylint==2.3.1
PyMySQL==0.9.3
pyOpenSSL==19.0.0
pytz==2019.1
pywin32==224
queuelib==1.5.0
requests==2.21.0
Scrapy==1.6.0
selenium==3.141.0
service-identity==18.1.0
six==1.12.0
sqlparse==0.3.0
tesseract==0.1.3
tesserocr==2.3.1
traitlets==4.3.2
Twisted==19.2.0
typed-ast==1.3.5
urllib3==1.24.3
w3lib==1.20.0
wcwidth==0.1.7
wrapt==1.11.1
zope.interface==4.6.0
